requirements.txt
# FarmTech UP - Dependencies
google-generativeai>=0.3.0
gitpython>=3.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
